electron-react-boilerplate/stylegan2_pytorch/requirements.txt
numpy
pillow
pyyaml
requests
scipy
tensorboard
torch
torchvision
tqdm
python-socketio
websocket-client